Refactor and Fetch data dynamically
- Separate DataUI from Landing page of the `Data` page;
- Add fetching of projects to show the data;
- Add selection of project from which the data is shown;
- Add filter by range of dates.
