+++ src/lib/services/data/index.ts
@@ -2,15 +2,15 @@ import { TRPCError } from '@trpc/server';
 
 import prisma from 'lib/db/prisma';
 
-import type { getDataDashboardProjectByLabelSchema } from './schemas';
+import type { getDataDashboardProjectByIdSchema } from './schemas';
 import type { z } from 'zod';
 
 export const getDataDashboardProject = async (
-  label: z.infer<typeof getDataDashboardProjectByLabelSchema>
+  id: z.infer<typeof getDataDashboardProjectByIdSchema>
 ) => {
   const request = await prisma.dataDashboardProject.findFirst({
     where: {
-      label,
+      id,
     },
     include: {
       data: {
@@ -20,6 +20,16 @@ export const getDataDashboardProject = async (
       },
     },
   });
+
+  if (!request) {
+    throw new TRPCError({ code: 'NOT_FOUND' });
+  }
+
+  return request;
+};
+
+export const getDataDashboardProjects = async () => {
+  const request = await prisma.dataDashboardProject.findMany();
 
   if (!request) {
     throw new TRPCError({ code: 'NOT_FOUND' });
